Added OrFail methods (`accountOrFail` and `budgetOrFail`) and some other changes
This fetches data if available or throws an exception that returns a not found response if data not available. This improves code reusability by a lot.

Changed method for fetching budget to join the `accounts` and `budgets` tables and query the `user_id` and `account_id`, This is more efficient than a seperate query to validate the account and another to fetch the budget from the database.