fix: Removed inefficient method for fetching accounts
This removes fetching accounts in a tree like structure, This includes recursing until all child accounts are found.
This is very problematic as if the number of accounts tied to an account is a large one it would attempt to fetch all of them.

Now it only fetches first level accounts, if child accounts are needed a new request has to be made to fetch the next level of accounts.